Wrapper function multi_rbh_overlap_based or construct_multi_rbh_correlation_based

Wrapper function multi_rbh_overlap_based or construct_multi_rbh_correlation_based

Usage

construct_multi_study_rbh(
  network_membership_list,
  method = c("overlap", "pearson", "spearman"),
  top_n = 50,
  memb_cut = 0,
  lower_quant = 0,
  upper_quant = 1,
  max_rank = 1,
  abs = FALSE,
  sparse = FALSE,
  binary = FALSE
)

Arguments

network_membership_list

a list containing community membership scores for each network. Where rownames contain unique gene ids and column

method

overlap, pearson, or spearman

top_n

the number of top genes based on membership scores with higher scores indicating stronger membership in a community

memb_cut

membership threshold for stricter thresholding. Only genes with membership score greater the threshold are used

lower_quant

indicates the quantile for the minimum correlation for the reciprocal best hits we will find.

upper_quant

indicates the quantile for correlations above which ANY metagene pairing will be considered a "hit."

max_rank

represents highest column and row rankings accepted for our reciprocal best hits network. Pure reciprocal best hits uses max_rank of 1.

abs

logical, take absolute values of correlations?

sparse

logical, use a sparse matrix to store network?

binary

logical, indicates whether or not the meta residual best hits matrix should show correlations or simply binary (as.numeric(correlation > 0)) output

Value

reciprocal best hits matrix between all metagenes across all of the datasets in the specified files, with dimensions named uniquely based on the file column names as well as the file names (to ensure uniqueness).

Details

top_n and memb_cut arguments are used if method = "overlap" and other arguments are used if method is "pearson" or "spearman".